rename helper folder and adding helper class add tests in components
