+++ libs/packages/components/src/lib/footer/footer.component.spec.ts
@@ -2,7 +2,7 @@
 import { async, ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
 import { SdsFooterComponent } from './footer.component';
 import { RouterTestingModule } from '@angular/router/testing';
-
+import { NavigationMode } from '../common-navigation/common-navigation-model';
 
 
 describe('SdsFooterComponent', () => {
@@ -24,6 +24,13 @@ describe('SdsFooterComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('event click', () => {
+    let navItem = { mode: NavigationMode.EVENT, text: 'test', route: '/' };
+    spyOn(component.linkEvent, 'emit');
+    component.linkClickEvent(navItem);
+    expect(component.linkEvent.emit).toHaveBeenCalledWith(navItem);
   });
 });
