+++ libs/packages/components/src/lib/footer/model/FooterModel.ts
@@ -1,4 +1,4 @@
-import { INavigationLink, NavigationMode } from '../../common-navigation-model/common-navigation-model';
+import { INavigationLink, NavigationMode } from '../../common-navigation/common-navigation-model';
 
 export class FooterModel {
