+++ libs/packages/components/src/lib/public-api.ts
@@ -9,6 +9,6 @@ export * from './side-navigation/side-navigation.module';
 export * from './search-result-list/search-result-list.module';
 export * from './footer/model/FooterModel';
 export * from './header/model/HeaderModel';
-export * from './common-navigation-model/common-navigation-model';
+export * from './common-navigation/common-navigation-model';
 export * from './side-navigation/model/side-navigation-model';
