+++ libs/packages/components/src/lib/side-navigation/model/side-navigation-model.ts
@@ -1,4 +1,4 @@
-import { INavigationLink, NavigationMode, Selectable } from '../../common-navigation-model/common-navigation-model';
+import { INavigationLink, NavigationMode, Selectable } from '../../common-navigation/common-navigation-model';
 
 export class SideNavigationModel {
